backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
passlib[bcrypt]==1.7.4
python-jose[cryptography]
PyJWT>=2.0.0
python-multipart>=0.0.5
